let libexec/scrape use proper default output dir
The bin/ is what sets NODE_BUILD_DEFINITIONS properly (prepending
share/node-build and appending all the plugins' share/node-build dirs)
so libexec/scrape was defaulting to empty string. Now we default
libexec/scrape to use share/node-build as well, so it can be executed
directly and still work as expected.