Better messaging for skipped defs
If a package is missing _both_ source tarball and has no applicable
binaries, throw a more descriptive error.

If catching an error during writing, print that error for details.
(like, the fact that it was skipped due to missing both source +
binaries)

Lastly, remove the warning message during parsing about missing source
package. Better to allow that printed at the time it's skipped.